[DSPDC-1699] Refactor and rename stage_data to load_hca
